Consolidate creation of nomad clients and backends
By injecting the nomad client and backends, it will make the APIs
simpler and easier to test. It is also more efficient as no longer
are we parser the config file multiple times.

A downside to this is, if the config is changed during running it
would display the updated config in the API calls. However, despite
this, the new jobs would not have been reloaded, leading to
inconsistencies.